services/control-api/requirements.txt
fastapi==0.115.6
uvicorn==0.34.0
pydantic==2.10.5
httpx==0.28.1
pytest==8.3.4
google-genai>=1.0.0